requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
commonmark==0.9.1
coverage==5.1
dictdiffer==0.8.1
flask==1.1.2
idna==2.9
importlib-metadata==1.6.0 ; python_version < '3.8'
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
more-itertools==8.2.0
multidict==4.7.5
notion==0.0.25
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyee==7.0.1
pyparsing==2.4.7
pytest-cov==2.8.1
pytest==5.4.1
python-slugify==4.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
soupsieve==2.0
text-unidecode==1.3
tzlocal==2.0.0
urllib3==1.25.9
wcwidth==0.1.9
werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0
